Added ProjectileHitActivator (WIP) "Replaced" deprecated Bukkit.getPlayerExact(String) Finalize some variables Refactoring (I know, you was waiting for it) Added player_level placeholder Fixed error-spam because of onSignClick Added some TODOs for future updates

+++ core/src/main/java/me/fromgate/reactions/activators/DropActivator.java
@@ -38,7 +38,7 @@ public class DropActivator extends Activator {
 		Variables.setTempVar("pickupDelay", Double.toString(de.getPickupDelay()));
 		boolean result = Actions.executeActivator(de.getPlayer(), this);
 		String pickupDelayStr = Variables.getTempVar("pickupDelay");
-		if (Util.FLOAT.matcher(pickupDelayStr).matches()) de.setPickupDelay(Double.parseDouble(pickupDelayStr));
+		if (Util.INT.matcher(pickupDelayStr).matches()) de.setPickupDelay(Integer.parseInt(pickupDelayStr));
 		Param itemParam = new Param(Variables.getTempVar("item"));
 		if (!itemParam.isEmpty()) {
 			String itemType = itemParam.getParam("type", "0");
