+++ core/src/main/java/me/fromgate/reactions/storage/DropStorage.java
@@ -10,9 +10,9 @@ import org.bukkit.inventory.ItemStack;
  */
 public class DropStorage extends RAStorage {
 	private ItemStack itemStack;
-	private Double pickupDelay;
+	private int pickupDelay;
 
-	public DropStorage(Player p, Item item, double pickupDelay) {
+	public DropStorage(Player p, Item item, int pickupDelay) {
 		super(p, ActivatorType.DROP);
 		this.itemStack = item.getItemStack();
 		this.pickupDelay = pickupDelay;
@@ -26,11 +26,11 @@ public class DropStorage extends RAStorage {
 		this.itemStack = itemStack;
 	}
 
-	public Double getPickupDelay() {
+	public int getPickupDelay() {
 		return this.pickupDelay;
 	}
 
-	public void setPickupDelay(Double pickupDelay) {
+	public void setPickupDelay(int pickupDelay) {
 		this.pickupDelay = pickupDelay;
 	}
